+++ Changelog.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Fixed 
+- Fix `ForbidJavaFilesTask` configuration failure when registered in non-android module with android module parent
 
 ## [0.12.0] - 2020-05-16
 ## Added